Refactored int test framework to include common class library
