+++ src/Monai.Deploy.WorkflowManager.sln
@@ -1,22 +1,4 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-
-<!--
-  ~ Copyright 2022 MONAI Consortium
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
--->
-
-Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32328.378
 MinimumVisualStudioVersion = 10.0.40219.1
@@ -91,6 +73,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.MonaiBackgroundService", "WorkflowManager\MonaiBackgroundService\Monai.Deploy.WorkflowManager.MonaiBackgroundService.csproj", "{93F4FE97-120C-44DC-9B21-69FB7EAB5846}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.MonaiBackgroundService.Tests", "..\tests\UnitTests\MonaiBackgroundService.Tests\Monai.Deploy.WorkflowManager.MonaiBackgroundService.Tests.csproj", "{EFECF826-B036-4689-B223-D791CD2C0F10}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monai.Deploy.WorkflowManager.IntegrationTests", "Shared\Test\Monai.Deploy.WorkflowManager.IntegrationTests\Monai.Deploy.WorkflowManager.IntegrationTests.csproj", "{B71E1EAF-74BB-4B69-B2BA-BC390232CFAC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -222,6 +206,10 @@ Global
 		{EFECF826-B036-4689-B223-D791CD2C0F10}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EFECF826-B036-4689-B223-D791CD2C0F10}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EFECF826-B036-4689-B223-D791CD2C0F10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B71E1EAF-74BB-4B69-B2BA-BC390232CFAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B71E1EAF-74BB-4B69-B2BA-BC390232CFAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B71E1EAF-74BB-4B69-B2BA-BC390232CFAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B71E1EAF-74BB-4B69-B2BA-BC390232CFAC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -259,6 +247,7 @@ Global
 		{99A847CA-7444-49BD-8814-C7CB68A0392F} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{93F4FE97-120C-44DC-9B21-69FB7EAB5846} = {AC5B198A-C3F2-4AD2-B532-E71B4630BDD5}
 		{EFECF826-B036-4689-B223-D791CD2C0F10} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
+		{B71E1EAF-74BB-4B69-B2BA-BC390232CFAC} = {37A19144-CEA5-47A2-9FFD-22C522E8B895}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DC0D56C8-D8CB-45CE-B528-F3DCF86D63ED}
