+++ src/Shared/Test/Monai.Deploy.WorkflowManager.IntegrationTests/Clients/RabbitConsumer.cs
@@ -18,7 +18,7 @@ using System.Text;
 using Newtonsoft.Json;
 using RabbitMQ.Client;
 
-namespace Monai.Deploy.WorkflowManager.IntegrationTests.Support
+namespace Monai.Deploy.WorkflowManager.IntegrationTests
 {
     public class RabbitConsumer
     {
