+++ src/Shared/Test/Monai.Deploy.WorkflowManager.IntegrationTests/Monai.Deploy.WorkflowManager.IntegrationTests.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Minio" Version="4.0.4" />
+    <PackageReference Include="MongoDB.Driver" Version="2.17.0" />
+    <PackageReference Include="Polly" Version="7.2.3" />
+    <PackageReference Include="RabbitMQ.Client" Version="6.4.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\TaskManager\API\Monai.Deploy.WorkflowManager.TaskManager.API.csproj" />
+    <ProjectReference Include="..\..\..\WorkflowManager\Contracts\Monai.Deploy.WorkflowManager.Contracts.csproj" />
+  </ItemGroup>
+
+</Project>
