--- tests/IntegrationTests/WorkflowExecutor.IntegrationTests/Support/MessageReceiver.cs
@@ -1,37 +1,0 @@
-/*
- * Copyright 2021-2022 MONAI Consortium
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-using RabbitMQ.Client;
-
-namespace Monai.Deploy.WorkflowManager.IntegrationTests.Support
-{
-    [Binding]
-    public class MessageReceiver : DefaultBasicConsumer
-    {
-        private readonly IModel _channel;
-
-        public MessageReceiver(IModel channel)
-        {
-            _channel = channel;
-        }
-
-        public override void HandleBasicDeliver(string consumerTag, ulong deliveryTag, bool redelivered, string exchange, string routingKey, IBasicProperties properties, ReadOnlyMemory<byte> body)
-        {
-            // TO DO Assertions
-            _channel.BasicAck(deliveryTag, false);
-        }
-    }
-}
